loanAgreement test cases

+++ com.cg.homeloan/src/main/java/com/cg/homeloan/entities/LoanApplication.java
@@ -72,7 +72,7 @@ public class LoanApplication {
 		this.status = Status.WAITING_FOR_LAND_VERIFICATION_OFFICE_APPROVAL;
 	}
 
-	public long getApplicationId() {
+	public int getApplicationId() {
 		return applicationId;
 	}
